OK pour calendrier !

+++ apps/back/src/app/calendar/days/days.service.ts
@@ -1,5 +1,5 @@
-import { Injectable } from '@nestjs/common';
-import { Days, EDays, ItemBase } from '@shared-interfaces/items';
+import { Injectable, NotFoundException } from '@nestjs/common';
+import { Days, EDays, ItemBase, ItemType } from '@shared-interfaces/items';
 import { InjectRepository } from '@nestjs/typeorm';
 import { DaysEntity } from './days.entity';
 import { Repository } from 'typeorm';
@@ -16,7 +16,7 @@ export class DaysService {
     const queryRecette: ItemBaseWithSlug[] = await this._daysEntityRepository
       .createQueryBuilder('days')
       .select(
-        'recette.articlesList, recette.id, recette.label, recette.description, recette.url, recette.tags, recette.itemType, days.slug, days.id as "tableIdentifier"'
+        'recette.articlesList, recette.id, recette.label, recette.description, recette.url, recette.tags, recette.itemType, days.slug, days.id as "tableIdentifier"',
       )
       .leftJoin('days.recetteId', 'recette')
       .where('days.recetteId is not null')
@@ -24,7 +24,9 @@ export class DaysService {
 
     const queryArticle: ItemBaseWithSlug[] = await this._daysEntityRepository
       .createQueryBuilder('days')
-      .select('article.id, article.label, article.description, article.url, article.tags, article.itemType, days.slug, days.id as "tableIdentifier"')
+      .select(
+        'article.id, article.label, article.description, article.url, article.tags, article.itemType, days.slug, days.id as "tableIdentifier"',
+      )
       .leftJoin('days.articleId', 'article')
       .where('days.articleId is not null')
       .getRawMany();
@@ -35,17 +37,24 @@ export class DaysService {
   }
 
   async putCalendarDay(days: DaysDto): Promise<void> {
-    // const diversType = divers.type === ItemType.RECETTE ? { recetteId: divers.itemId } : { articleId: divers.itemId };
-    // const entity = this._diversEntityRepository.create({ ...diversType });
-    // if (!entity) {
-    //   throw new NotFoundException();
-    // }
-    // console.log(entity);
-    // await this._diversEntityRepository.save(entity);
+    console.log(days);
+    const daysType =
+      days.type === ItemType.RECETTE
+        ? { recetteId: days.itemId, slug: days.slug }
+        : { articleId: days.itemId, slug: days.slug };
+    const entity = this._daysEntityRepository.create({ ...daysType });
+    if (!entity) {
+      throw new NotFoundException();
+    }
+    console.log(entity);
+    await this._daysEntityRepository.save(entity);
   }
 
   async deleteCalendarDay(id: number): Promise<void> {
     const entity = await this._daysEntityRepository.findOneBy({ id });
+    if (!entity) {
+      throw new NotFoundException();
+    }
     await this._daysEntityRepository.remove(entity);
   }
 }
